requirements.txt
faker
tqdm
psycopg2-binary
mysql-connector-python
numpy
pandas
